@@ -6,20 +6,22 @@
 
     <groupId>com.smockin</groupId>
     <artifactId>smockin</artifactId>
-    <version>2.20.3</version>
+    <version>2.20.4</version>
     <packaging>jar</packaging>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.11</version>
+        <version>3.2.5</version>
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>11</java.version>
-        <spring-boot-test.version>3.0.8</spring-boot-test.version>
+        <java.version>17</java.version>
+        <!-- Override Jetty version to be compatible with Javalin 6.x -->
+        <jetty.version>12.0.14</jetty.version>
+        <spring-boot-test.version>3.2.5</spring-boot-test.version>
         <spark-core.version>2.9.4</spark-core.version>
         <h2.version>2.3.232</h2.version>
         <hikaricp.version>2.6.1</hikaricp.version>
