@@ -6,6 +6,7 @@ import com.smockin.admin.service.UserKeyValueDataService;
 import com.smockin.mockserver.exception.InboundParamMatchException;
 import com.smockin.mockserver.service.enums.ParamMatchTypeEnum;
 import com.smockin.utils.GeneralUtils;
+import io.javalin.http.Context;
 import org.apache.commons.lang3.RandomUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
@@ -13,7 +14,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import spark.Request;
 
 import java.text.SimpleDateFormat;
 
@@ -35,7 +35,7 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
     private static final String GENERAL_ERROR = "Error processing inbound param matching. Please check your token syntax";
 
     @Override
-    public String enrichWithInboundParamMatches(final Request req,
+    public String enrichWithInboundParamMatches(final Context ctx,
                                                 final String mockPath,
                                                 final String responseBody,
                                                 final String userCtxPath,
@@ -45,7 +45,7 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
             return null;
         }
 
-        final String sanitizedUserCtxInboundPath = GeneralUtils.sanitizeMultiUserPath(smockinUserService.getUserMode(), req.pathInfo(), userCtxPath);
+        final String sanitizedUserCtxInboundPath = GeneralUtils.sanitizeMultiUserPath(smockinUserService.getUserMode(), ctx.path(), userCtxPath);
 
         String enrichedResponseBody = responseBody;
 
@@ -62,7 +62,7 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
             final String r;
 
             try {
-                r = processParamMatch(req, mockPath, enrichedResponseBody, sanitizedUserCtxInboundPath, mockOwnerUserId);
+                r = processParamMatch(ctx, mockPath, enrichedResponseBody, sanitizedUserCtxInboundPath, mockOwnerUserId);
             } catch (Throwable ex) {
                 logger.error(ex.getMessage());
                 throw new InboundParamMatchException(GENERAL_ERROR);
@@ -80,7 +80,7 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
         return enrichedResponseBody;
     }
 
-    String processParamMatch(final Request req,
+    String processParamMatch(final Context ctx,
                              final String mockPath,
                              final String responseBody,
                              final String sanitizedUserCtxInboundPath,
@@ -99,15 +99,15 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
 
         // Determine the matching token type, is it a requestHeader, requestParameter, pathVar, etc...
         if (ParamMatchTypeEnum.lookUpKvp.equals(paramMatchType)) {
-            return processKvp(matchStartingPosition, sanitizedUserCtxInboundPath, mockPath, req, responseBody, mockOwnerUserId);
+            return processKvp(matchStartingPosition, sanitizedUserCtxInboundPath, mockPath, ctx, responseBody, mockOwnerUserId);
         }
 
         if (ParamMatchTypeEnum.requestHeader.equals(paramMatchType)) {
-            return processRequestHeader(matchStartingPosition, req, responseBody);
+            return processRequestHeader(matchStartingPosition, ctx, responseBody);
         }
 
         if (ParamMatchTypeEnum.requestParameter.equals(paramMatchType)) {
-            return processRequestParameter(matchStartingPosition, req, responseBody);
+            return processRequestParameter(matchStartingPosition, ctx, responseBody);
         }
 
         if (ParamMatchTypeEnum.pathVar.equals(paramMatchType)) {
@@ -115,9 +115,10 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
         }
 
         if (ParamMatchTypeEnum.requestBody.equals(paramMatchType)) {
+            final String body = ctx.body();
             return StringUtils.replaceIgnoreCase(responseBody,
                     ParamMatchTypeEnum.PARAM_PREFIX + ParamMatchTypeEnum.requestBody,
-                    (req.body() != null) ? req.body() : "",
+                    (body != null) ? body : "",
                     1);
         }
 
@@ -183,7 +184,7 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
     String processKvp(final int matchStartingPosition,
                       final String sanitizedUserCtxInboundPath,
                       final String mockPath,
-                      final Request req,
+                      final Context ctx,
                       final String responseBody,
                       final long mockOwnerUserId) {
 
@@ -197,8 +198,8 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
         }
 
         if (logger.isDebugEnabled()) {
-            logger.debug("RAW KVP: " + kvpKey);
-            logger.debug("Cleaned KVP : " + sanitisedKvpKey);
+            logger.debug("RAW KVP: {}" , kvpKey);
+            logger.debug("Cleaned KVP : {}" , sanitisedKvpKey);
         }
 
         // Check if kvpKey is a nested ParamMatchTypeEnum itself
@@ -208,32 +209,22 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
         if (isNested) {
 
             if (logger.isDebugEnabled()) {
-                logger.debug("Nested KVP request type: " + kvpMatchResult.getLeft());
+                logger.debug("Nested KVP request type: {}" , kvpMatchResult.getLeft());
             }
 
             final String nestedRequestKey = extractArgName(kvpMatchResult.getRight(), kvpMatchResult.getLeft(), sanitisedKvpKey, isNested);
 
             if (logger.isDebugEnabled()) {
-                logger.debug("Nested KVP request key: " + nestedRequestKey);
+                logger.debug("Nested KVP request key: {}" , nestedRequestKey);
             }
-
-            switch (kvpMatchResult.getLeft()) {
-                case requestHeader:
-                    sanitisedKvpKey = GeneralUtils.findHeaderIgnoreCase(req, sanitiseArgName(nestedRequestKey));
-                    break;
-                case requestParameter:
-                    sanitisedKvpKey = GeneralUtils.extractRequestParamByName(req, sanitiseArgName(nestedRequestKey));
-                    break;
-                case pathVar:
-                    sanitisedKvpKey = GeneralUtils.findPathVarIgnoreCase(sanitizedUserCtxInboundPath, mockPath, sanitiseArgName(nestedRequestKey));
-                    break;
-                case requestBody:
-                    sanitisedKvpKey = req.body();
-                    break;
-                default:
-                    sanitisedKvpKey = null;
-                    break;
-            }
+            
+            sanitisedKvpKey = switch (kvpMatchResult.getLeft()) {
+                case requestHeader -> GeneralUtils.findHeaderIgnoreCase(ctx, sanitiseArgName(nestedRequestKey));
+                case requestParameter -> GeneralUtils.extractRequestParamByName(ctx, sanitiseArgName(nestedRequestKey));
+                case pathVar -> GeneralUtils.findPathVarIgnoreCase(sanitizedUserCtxInboundPath, mockPath, sanitiseArgName(nestedRequestKey));
+                case requestBody -> ctx.body();
+                default -> null;
+            };
         }
 
         final UserKeyValueDataDTO userKeyValueDataDTO = (sanitisedKvpKey != null)
@@ -250,10 +241,10 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
                 1);
     }
 
-    String processRequestHeader(final int matchStartingPosition, final Request req, final String responseBody) {
+    String processRequestHeader(final int matchStartingPosition, final Context ctx, final String responseBody) {
 
         final String headerName = extractArgName(matchStartingPosition, ParamMatchTypeEnum.requestHeader, responseBody, false);
-        final String headerValue = GeneralUtils.findHeaderIgnoreCase(req, sanitiseArgName(headerName));
+        final String headerValue = GeneralUtils.findHeaderIgnoreCase(ctx, sanitiseArgName(headerName));
 
         if (logger.isDebugEnabled()) {
             logger.debug("raw header: " + headerName);
@@ -268,16 +259,16 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
     }
 
     String processRequestParameter(final int matchStartingPosition,
-                                   final Request req,
+                                   final Context ctx,
                                    final String responseBody) {
 
         final String requestParamName = extractArgName(matchStartingPosition, ParamMatchTypeEnum.requestParameter, responseBody, false);
-        final String requestParamValue = GeneralUtils.extractRequestParamByName(req, sanitiseArgName(requestParamName));
+        final String requestParamValue = GeneralUtils.extractRequestParamByName(ctx, sanitiseArgName(requestParamName));
 
         if (logger.isDebugEnabled()) {
-            logger.debug("RAW request param: " + requestParamName);
-            logger.debug("Cleaned request param: " + sanitiseArgName(requestParamName));
-            logger.debug("Request param value: " + requestParamValue);
+            logger.debug("RAW request param: {}" , requestParamName);
+            logger.debug("Cleaned request param: {}" , sanitiseArgName(requestParamName));
+            logger.debug("Request param value: {}" , requestParamValue);
         }
 
         return StringUtils.replaceIgnoreCase(responseBody,
@@ -293,9 +284,9 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
         final String pathVariableValue = GeneralUtils.findPathVarIgnoreCase(sanitizedUserCtxInboundPath, mockPath, sanitiseArgName(pathVariableName));
 
         if (logger.isDebugEnabled()) {
-            logger.debug("RAW path var: " + pathVariableName);
-            logger.debug("Cleaned path var : " + sanitiseArgName(pathVariableName));
-            logger.debug("Path var value: " + pathVariableValue);
+            logger.debug("RAW path var: {}" , pathVariableName);
+            logger.debug("Cleaned path var : {}" , sanitiseArgName(pathVariableName));
+            logger.debug("Path var value: {}" , pathVariableValue);
         }
 
         return StringUtils.replaceIgnoreCase(responseBody,
@@ -310,7 +301,7 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
         final String randomNumberContent = extractArgName(matchStartingPosition, ParamMatchTypeEnum.randomNumber, responseBody, false);
 
         if (logger.isDebugEnabled()) {
-            logger.debug("Random number params: " + randomNumberContent);
+            logger.debug("Random number params: {}" , randomNumberContent);
         }
 
         if (randomNumberContent == null) {
@@ -332,7 +323,7 @@ public class InboundParamMatchServiceImpl implements InboundParamMatchService {
         final int randomValue = RandomUtils.nextInt(startInc, endExcl);
 
         if (logger.isDebugEnabled()) {
-            logger.debug("Random number value: " + randomValue);
+            logger.debug("Random number value: {}", randomValue);
         }
 
         return StringUtils.replaceIgnoreCase(responseBody,
