@@ -4,31 +4,34 @@ import com.smockin.admin.enums.UserModeEnum;
 import com.smockin.admin.service.AuthService;
 import com.smockin.admin.service.SmockinUserService;
 import com.smockin.utils.GeneralUtils;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
+import org.jspecify.annotations.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
+import org.springframework.web.servlet.HandlerInterceptor;
+
 import javax.annotation.PostConstruct;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 @Component
-public class AuthInterceptor extends HandlerInterceptorAdapter {
+@RequiredArgsConstructor
+public class AuthInterceptor implements HandlerInterceptor {
 
     private final Logger logger = LoggerFactory.getLogger(AuthInterceptor.class);
 
-    @Autowired
-    private SmockinUserService smockinUserService;
 
-    @Autowired
-    private AuthService authService;
+    private final SmockinUserService smockinUserService;
+
+   
+    private final AuthService authService;
 
     // Resorted to using json, in place of defining the list structure in yaml, due to a bug introduced
     // in Spring Boot v2 where forward slashes (& other characters) seem to be stripped out when injected in.
@@ -43,7 +46,7 @@ public class AuthInterceptor extends HandlerInterceptorAdapter {
             http://www.baeldung.com/spring-mvc-handlerinterceptor
     */
     @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+    public boolean preHandle(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response, @NonNull Object handler) throws Exception {
         logger.debug("Request received...");
 
         if (!UserModeEnum.ACTIVE.equals(smockinUserService.getUserMode())) {
@@ -81,7 +84,7 @@ public class AuthInterceptor extends HandlerInterceptorAdapter {
         final int wildCardPathPos = exclusionKey.indexOf("*");
 
         if (wildCardFilePos > -1) {
-            return inboundUrl.endsWith(exclusionKey.substring(wildCardFilePos + 1, exclusionKey.length()));
+            return inboundUrl.endsWith(exclusionKey.substring(wildCardFilePos + 1));
         } else if (wildCardPathPos > -1) {
             return inboundUrl.startsWith(exclusionKey.substring(0, wildCardPathPos));
         }
@@ -101,8 +104,7 @@ public class AuthInterceptor extends HandlerInterceptorAdapter {
 
         if (logger.isDebugEnabled()) {
             logger.debug("Current Exclusions:");
-            exclusions.entrySet()
-                    .forEach(e -> logger.debug("exclusion path: " + e.getKey() + ", method: " + e.getValue()));
+            exclusions.forEach((key, value) -> logger.debug("exclusion path: {} , method: {}" ,key, value));
         }
 
     }
@@ -110,8 +112,8 @@ public class AuthInterceptor extends HandlerInterceptorAdapter {
     private void debugRequest(final HttpServletRequest request) {
 
         if (logger.isDebugEnabled()) {
-            logger.debug("Request URI: " + request.getRequestURI());
-            logger.debug("Request Method: " + request.getMethod());
+            logger.debug("Request URI: {}" , request.getRequestURI());
+            logger.debug("Request Method: {}" , request.getMethod());
         }
 
     }
