@@ -1,8 +1,12 @@
 package com.smockin.admin.persistence.entity;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Data;
 import org.hibernate.annotations.ColumnDefault;
-import javax.persistence.*;
 
 /**
  * Created by mgallina.
