@@ -1,10 +1,16 @@
 package com.smockin.admin.persistence.entity;
 
 import com.smockin.admin.persistence.enums.ServerTypeEnum;
+import jakarta.persistence.Column;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Table;
 import lombok.Data;
 import org.hibernate.annotations.ColumnDefault;
 
-import javax.persistence.*;
 import java.util.HashMap;
 import java.util.Map;
 
