@@ -7,9 +7,6 @@ import com.smockin.admin.service.utils.UserTokenServiceUtils;
 import com.smockin.utils.GeneralUtils;
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.eclipse.jetty.websocket.api.WebSocketBehavior;
-import org.eclipse.jetty.websocket.api.WebSocketPolicy;
-import org.eclipse.jetty.websocket.server.WebSocketServerFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,10 +19,8 @@ import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
 import org.springframework.web.socket.handler.TextWebSocketHandler;
 import org.springframework.web.socket.server.HandshakeInterceptor;
-import org.springframework.web.socket.server.jetty.JettyRequestUpgradeStrategy;
 import org.springframework.web.socket.server.support.DefaultHandshakeHandler;
 
-import javax.servlet.ServletContext;
 import java.util.Map;
 
 @Configuration
@@ -35,9 +30,6 @@ public class WebSocketConfig implements WebSocketConfigurer {
     private final Logger logger = LoggerFactory.getLogger(WebSocketConfig.class);
 
     private static final String URL = "/liveLoggingFeed/*/*";
-
-    @Autowired
-    private ServletContext servletContext;
 
     @Autowired
     private LiveLoggingHandler mockLogFeedHandler;
@@ -53,14 +45,7 @@ public class WebSocketConfig implements WebSocketConfigurer {
     public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
         registry.addHandler(((TextWebSocketHandler) mockLogFeedHandler), URL)
                 .addInterceptors(userInterceptor())
-                .setHandshakeHandler(handshakeHandler());
-    }
-
-    private DefaultHandshakeHandler handshakeHandler() {
-        return new DefaultHandshakeHandler(
-                new JettyRequestUpgradeStrategy(
-                        new WebSocketServerFactory(servletContext,
-                                new WebSocketPolicy(WebSocketBehavior.SERVER))));
+                .setHandshakeHandler(new DefaultHandshakeHandler());
     }
 
     public HandshakeInterceptor userInterceptor() {
