@@ -5,13 +5,29 @@ import com.smockin.admin.exception.ValidationException;
 import com.smockin.admin.persistence.enums.RestMethodEnum;
 import com.smockin.admin.service.SmockinUserService;
 import com.smockin.admin.websocket.LiveLoggingHandler;
-import com.smockin.mockserver.dto.*;
+import com.smockin.mockserver.dto.BlockedPathToRelease;
+import com.smockin.mockserver.dto.LiveBlockPath;
+import com.smockin.mockserver.dto.LiveloggingUserOverrideResponse;
+import com.smockin.mockserver.dto.MockServerState;
+import com.smockin.mockserver.dto.MockedServerConfigDTO;
+import com.smockin.mockserver.dto.ProxyForwardConfigCacheDTO;
 import com.smockin.mockserver.exception.MockServerException;
-import com.smockin.mockserver.service.*;
-import com.smockin.mockserver.service.ws.SparkWebSocketEchoService;
+import com.smockin.mockserver.service.HttpProxyService;
+import com.smockin.mockserver.service.InboundParamMatchService;
+import com.smockin.mockserver.service.MockOrderingCounterService;
+import com.smockin.mockserver.service.RuleEngine;
+import com.smockin.mockserver.service.ServerSideEventService;
+import com.smockin.mockserver.service.WebSocketService;
+import com.smockin.mockserver.service.enums.WebSocketCommandEnum;
 import com.smockin.utils.GeneralUtils;
 import com.smockin.utils.LiveLoggingUtils;
+import io.javalin.Javalin;
+import io.javalin.config.RoutesConfig;
+import io.javalin.http.Context;
+import jakarta.servlet.http.HttpServletResponse;
 import org.apache.commons.lang3.StringUtils;
+import org.eclipse.jetty.websocket.api.Callback;
+import org.eclipse.jetty.websocket.api.Session;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,12 +35,12 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import spark.Request;
-import spark.Response;
-import spark.Spark;
 
-import javax.servlet.http.HttpServletResponse;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -77,6 +93,9 @@ public class MockedRestServerEngine {
     private final Object serverStateMonitor = new Object();
     private MockServerState serverState = new MockServerState(false, 0);
 
+    // Javalin app instance
+    private Javalin app;
+
     // Live logging response blocker
     // TODO find a smarter way to do this (i.e BlockingQueue...)
     private final Object responseBlockingMonitor = new Object();
@@ -98,20 +117,10 @@ public class MockedRestServerEngine {
 
         updateProxyMode(config.isProxyMode());
 
-        // Define all web socket routes first as the Spark framework requires this
-        buildWebSocketEndpoints(isMultiUserMode);
-
-        // Handle Cross-Origin Resource Sharing (CORS) support
-        handleCORS(config);
-
         proxyMappingCache.init(allProxyForwardConfig);
 
-        // Next handle all HTTP RESTFul web service routes
-        buildGlobalHttpEndpointsHandler(isMultiUserMode);
-
-        applyTrafficLogging(config.isProxyMode());
-
-        initServer(config.getPort());
+        // In Javalin 7, all routes must be defined in the config block
+        initServer(config.getPort(), config, isMultiUserMode);
     }
 
     public MockServerState getCurrentState() throws MockServerException {
@@ -121,53 +130,158 @@ public class MockedRestServerEngine {
     }
 
     public void shutdown() throws MockServerException {
-
         try {
-
             serverSideEventService.interruptAndClearAllHeartBeatThreads();
-
-            Spark.stop();
-
-            // Having dug around the source code, 'Spark.stop()' runs off a different thread when stopping the server and removing it's state such as routes, etc.
-            // This means that calling 'Spark.port()' immediately after stop, results in an IllegalStateException, as the
-            // 'initialized' flag is checked in the current thread and is still marked as true.
-            // (The error thrown: java.lang.IllegalStateException: This must be done before route mapping has begun)
-            // Short of editing the Spark source to fix this, I have therefore had to add this hack to buy the 'stop' thread time to complete.
-            Thread.sleep(3000);
-
+            if (app != null) {
+                app.stop();
+            }
             synchronized (serverStateMonitor) {
                 serverState.setRunning(false);
             }
-
             clearState();
-
         } catch (Throwable ex) {
             throw new MockServerException(ex);
         }
-
     }
 
-    void initServer(final int port) throws MockServerException {
+    void initServer(final int port, final MockedServerConfigDTO config, final boolean isMultiUserMode) throws MockServerException {
         logger.debug("initServer called");
 
         try {
-
             clearState();
-
-            Spark.init();
-
-            // Blocks the current thread (using a CountDownLatch under the hood) until the server is fully initialised.
-            Spark.awaitInitialization();
-
+            if (app != null) {
+                app.stop();
+            }
+            
+            final boolean enableCors = Boolean.TRUE.toString().equalsIgnoreCase(
+                config.getNativeProperties().get(GeneralUtils.ENABLE_CORS_PARAM));
+            final boolean isProxyMode = config.isProxyMode();
+            
+            app = Javalin.create(javalinConfig -> {
+                // Configure routes in Javalin 7
+                configureRoutes(javalinConfig.routes, isMultiUserMode, enableCors, isProxyMode);
+            }).start(port);
+            
             synchronized (serverStateMonitor) {
                 serverState.setRunning(true);
                 serverState.setPort(port);
             }
-
         } catch (Throwable ex) {
             throw new MockServerException(ex);
         }
-
+    }
+    
+    private void configureRoutes(RoutesConfig routes, boolean isMultiUserMode, boolean enableCors, boolean isProxyMode) {
+        // WebSocket endpoints
+        routes.ws(SPARK_WILDCARD_PATH, ws -> {
+            ws.onConnect(ctx -> webSocketService.registerSession(ctx.session, isMultiUserMode));
+            ws.onClose(ctx -> webSocketService.removeSession(ctx.session));
+            ws.onMessage(ctx -> {
+                final String message = ctx.message();
+                final Session session = ctx.session;
+                if (WebSocketCommandEnum.SMOCKIN_ID.name().equals(message)) {
+                    session.sendText(webSocketService.getExternalId(session), Callback.NOOP);
+                    return;
+                }
+                webSocketService.respondToMessage(session, message);
+            });
+        });
+        
+        // CORS handling
+        if (enableCors) {
+            routes.options(GeneralUtils.PATH_WILDCARD, ctx -> {
+                final String accessControlRequestHeaders = ctx.header(HttpHeaders.ACCESS_CONTROL_REQUEST_HEADERS);
+                if (accessControlRequestHeaders != null) {
+                    ctx.header(HttpHeaders.ACCESS_CONTROL_ALLOW_HEADERS, accessControlRequestHeaders);
+                }
+                final String accessControlRequestMethod = ctx.header(HttpHeaders.ACCESS_CONTROL_REQUEST_METHOD);
+                if (accessControlRequestMethod != null) {
+                    ctx.header(HttpHeaders.ACCESS_CONTROL_ALLOW_METHODS, accessControlRequestMethod);
+                }
+                ctx.status(HttpStatus.OK.value());
+                ctx.result(HttpStatus.OK.name());
+            });
+            routes.beforeMatched(ctx -> ctx.header(HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN, GeneralUtils.PATH_WILDCARD));
+        }
+        
+        // HTTP endpoints
+        // HEAD
+        routes.head(GeneralUtils.PATH_WILDCARD, ctx -> {
+            processResponse(ctx, isMultiUserMode, proxyModeEnabled.get());
+            checkForAndHandleBlockSwapAndMock(ctx, isMultiUserMode, proxyModeEnabled.get());
+            ctx.result("");
+        });
+        // GET
+        routes.get(GeneralUtils.PATH_WILDCARD, ctx -> {
+            if (isWebSocketUpgradeRequest(ctx)) {
+                ctx.status(HttpStatus.OK.value());
+                return;
+            }
+            final String responseBody = processResponse(ctx, isMultiUserMode, proxyModeEnabled.get());
+            final Optional<String> amendmentOpt = checkForAndHandleBlockSwapAndMock(ctx, isMultiUserMode, proxyModeEnabled.get());
+            ctx.result(amendmentOpt.orElse(responseBody));
+        });
+        // POST
+        routes.post(GeneralUtils.PATH_WILDCARD, ctx -> {
+            final String responseBody = processResponse(ctx, isMultiUserMode, proxyModeEnabled.get());
+            final Optional<String> amendmentOpt = checkForAndHandleBlockSwapAndMock(ctx, isMultiUserMode, proxyModeEnabled.get());
+            ctx.result(amendmentOpt.orElse(responseBody));
+        });
+        // PUT
+        routes.put(GeneralUtils.PATH_WILDCARD, ctx -> {
+            final String responseBody = processResponse(ctx, isMultiUserMode, proxyModeEnabled.get());
+            final Optional<String> amendmentOpt = checkForAndHandleBlockSwapAndMock(ctx, isMultiUserMode, proxyModeEnabled.get());
+            ctx.result(amendmentOpt.orElse(responseBody));
+        });
+        // DELETE
+        routes.delete(GeneralUtils.PATH_WILDCARD, ctx -> {
+            final String responseBody = processResponse(ctx, isMultiUserMode, proxyModeEnabled.get());
+            final Optional<String> amendmentOpt = checkForAndHandleBlockSwapAndMock(ctx, isMultiUserMode, proxyModeEnabled.get());
+            ctx.result(amendmentOpt.orElse(responseBody));
+        });
+        // PATCH
+        routes.patch(GeneralUtils.PATH_WILDCARD, ctx -> {
+            final String responseBody = processResponse(ctx, isMultiUserMode, proxyModeEnabled.get());
+            final Optional<String> amendmentOpt = checkForAndHandleBlockSwapAndMock(ctx, isMultiUserMode, proxyModeEnabled.get());
+            ctx.result(amendmentOpt.orElse(responseBody));
+        });
+        
+        // Traffic logging
+        routes.beforeMatched(ctx -> {
+            if (ctx.header(webSocketService.WS_SEC_WEBSOCKET_KEY) != null) {
+                return;
+            }
+            final String traceId = GeneralUtils.generateUUID();
+            ctx.attribute(GeneralUtils.LOG_REQ_ID, traceId);
+            ctx.header(GeneralUtils.LOG_REQ_ID, traceId);
+            final Map<String, String> reqHeaders = ctx.headerMap();
+            reqHeaders.put(GeneralUtils.LOG_REQ_ID, traceId);
+            liveLoggingHandler.broadcast(
+                    LiveLoggingUtils.buildLiveLogInboundDTO(
+                        ctx.attribute(GeneralUtils.LOG_REQ_ID),
+                        ctx.method().name(),
+                        ctx.path(),
+                        reqHeaders,
+                        ctx.body(),
+                        isProxyMode,
+                        GeneralUtils.extractAllRequestParams(ctx)));
+        });
+        routes.afterMatched(ctx -> {
+            if (ctx.header(WebSocketService.WS_SEC_WEBSOCKET_KEY) != null
+                    || ServerSideEventService.SSE_EVENT_STREAM_HEADER.equals(ctx.header(HttpHeaders.CONTENT_TYPE))) {
+                return;
+            }
+            final Map<String, String> respHeaders = mockedRestServerEngineUtils.extractResponseHeadersAsMap(ctx);
+            respHeaders.put(GeneralUtils.LOG_REQ_ID, ctx.attribute(GeneralUtils.LOG_REQ_ID));
+            liveLoggingHandler.broadcast(
+                    LiveLoggingUtils.buildLiveLogOutboundDTO(
+                        ctx.attribute(GeneralUtils.LOG_REQ_ID),
+                        ctx.path(),
+                        ctx.status().getCode(),
+                        respHeaders,
+                        ctx.result(),
+                        isProxyMode));
+        });
     }
 
     void initServerConfig(final MockedServerConfigDTO config) {
@@ -175,173 +289,21 @@ public class MockedRestServerEngine {
 
         if (logger.isDebugEnabled())
             logger.debug(config.toString());
-
-        Spark.port(config.getPort());
-        Spark.threadPool(config.getMaxThreads(), config.getMinThreads(), config.getTimeOutMillis());
+        // Javalin config is handled in start/initServer
     }
 
-    void buildWebSocketEndpoints(final boolean isMultiUserMode) {
-
-        Spark.webSocket(SPARK_WILDCARD_PATH, new SparkWebSocketEchoService(webSocketService, isMultiUserMode));
-    }
-
-    private void applyTrafficLogging(final boolean isUsingProxyMode) {
-
-        // Live logging filter
-        Spark.before((request, response) -> {
-
-            if (request.raw().getHeader(webSocketService.WS_SEC_WEBSOCKET_KEY) != null) {
-                return;
-            }
-
-            final String traceId = GeneralUtils.generateUUID();
-
-            request.attribute(GeneralUtils.LOG_REQ_ID, traceId);
-            response.raw()
-                    .addHeader(GeneralUtils.LOG_REQ_ID, traceId);
-
-            final Map<String, String> reqHeaders = request
-                    .headers()
-                    .stream()
-                    .collect(Collectors.toMap(h -> h, h -> request.headers(h)));
-
-            reqHeaders.put(GeneralUtils.LOG_REQ_ID, traceId);
-
-            liveLoggingHandler.broadcast(
-                    LiveLoggingUtils.buildLiveLogInboundDTO(
-                        request.attribute(GeneralUtils.LOG_REQ_ID),
-                        request.requestMethod(),
-                        request.pathInfo(),
-                        reqHeaders,
-                        request.body(),
-                        isUsingProxyMode,
-                        GeneralUtils.extractAllRequestParams(request)));
-        });
-
-        Spark.afterAfter((request, response) -> {
-
-            if (request.raw().getHeader(webSocketService.WS_SEC_WEBSOCKET_KEY) != null
-                    || serverSideEventService.SSE_EVENT_STREAM_HEADER.equals(response.raw().getHeader(HttpHeaders.CONTENT_TYPE))) {
-                return;
-            }
-
-            final Map<String, String> respHeaders = mockedRestServerEngineUtils.extractResponseHeadersAsMap(response);
-
-            respHeaders.put(GeneralUtils.LOG_REQ_ID, request.attribute(GeneralUtils.LOG_REQ_ID));
-
-            liveLoggingHandler.broadcast(
-                    LiveLoggingUtils.buildLiveLogOutboundDTO(
-                        request.attribute(GeneralUtils.LOG_REQ_ID),
-                        request.pathInfo(),
-                        response.raw().getStatus(),
-                        respHeaders,
-                        response.body(),
-                        isUsingProxyMode));
-        });
-
-    }
-
-    void buildGlobalHttpEndpointsHandler(final boolean isMultiUserMode) {
-        logger.debug("buildGlobalHttpEndpointsHandler called");
-
-        // HEAD
-        Spark.head(GeneralUtils.PATH_WILDCARD, (request, response) -> {
-
-            processResponse(request, response, isMultiUserMode, proxyModeEnabled.get());
-            checkForAndHandleBlockSwapAndMock(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            return "";
-        });
-
-        // GET
-        Spark.get(GeneralUtils.PATH_WILDCARD, (request, response) -> {
-
-            if (isWebSocketUpgradeRequest(request)) {
-                response.status(HttpStatus.OK.value());
-                return null;
-            }
-
-            final String responseBody = processResponse(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            final Optional<String> amendmentOpt =
-                    checkForAndHandleBlockSwapAndMock(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            return (amendmentOpt.isPresent())
-                    ? amendmentOpt.get()
-                    : responseBody;
-        });
-
-        // POST
-        Spark.post(GeneralUtils.PATH_WILDCARD, (request, response) -> {
-
-            final String responseBody = processResponse(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            final Optional<String> amendmentOpt =
-                    checkForAndHandleBlockSwapAndMock(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            return (amendmentOpt.isPresent())
-                    ? amendmentOpt.get()
-                    : responseBody;
-        });
-
-        // PUT
-        Spark.put(GeneralUtils.PATH_WILDCARD, (request, response) -> {
-
-            final String responseBody = processResponse(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            final Optional<String> amendmentOpt =
-                    checkForAndHandleBlockSwapAndMock(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            return (amendmentOpt.isPresent())
-                    ? amendmentOpt.get()
-                    : responseBody;
-        });
-
-        // DELETE
-        Spark.delete(GeneralUtils.PATH_WILDCARD, (request, response) -> {
-
-            final String responseBody = processResponse(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            final Optional<String> amendmentOpt =
-                    checkForAndHandleBlockSwapAndMock(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            return (amendmentOpt.isPresent())
-                    ? amendmentOpt.get()
-                    : responseBody;
-        });
-
-        // PATCH
-        Spark.patch(GeneralUtils.PATH_WILDCARD, (request, response) -> {
-
-            final String responseBody = processResponse(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            final Optional<String> amendmentOpt =
-                    checkForAndHandleBlockSwapAndMock(request, response, isMultiUserMode, proxyModeEnabled.get());
-
-            return (amendmentOpt.isPresent())
-                    ? amendmentOpt.get()
-                    : responseBody;
-        });
-
-    }
-
-    String processResponse(final Request request,
-                   final Response response,
+    String processResponse(final Context ctx,
                    final boolean isMultiUserMode,
                    final boolean isProxyMode) {
-
-        return mockedRestServerEngineUtils.loadMockedResponse(request, response, isMultiUserMode, isProxyMode)
-                .orElseGet(() ->
-                        handleNotFoundResponse(response));
+        return mockedRestServerEngineUtils.loadMockedResponse(ctx, isMultiUserMode, isProxyMode)
+                .orElseGet(() -> handleNotFoundResponse(ctx));
     }
 
-    Optional<String> checkForAndHandleBlockSwapAndMock(final Request request,
-                                                       final Response response,
+    Optional<String> checkForAndHandleBlockSwapAndMock(final Context ctx,
                                                        final boolean isMultiUserMode,
                                                        final boolean isProxyMode)
             throws InterruptedException {
-
-        if (blockLoggingResponse(request, response, isProxyMode)) {
+        if (blockLoggingResponse(ctx, isProxyMode)) {
 
             logger.debug("Endpoint match made. Blocking response...");
 
@@ -352,7 +314,7 @@ public class MockedRestServerEngine {
                     // Wait for response amendment for this request (by traceId)
                     responseBlockingMonitor.wait();
 
-                    final String traceId = request.attribute(GeneralUtils.LOG_REQ_ID);
+                    final String traceId = ctx.attribute(GeneralUtils.LOG_REQ_ID);
 
                     // Release request if liveBlocking is disabled at any stage
                     if (!liveBlockingModeEnabled.get()) {
@@ -370,19 +332,19 @@ public class MockedRestServerEngine {
 
                                     if (p.getMethod().isPresent()) {
 
-                                        return request.requestMethod().equalsIgnoreCase(p.getMethod().get().name())
-                                                && StringUtils.equals(request.pathInfo(), p.getPathPattern());
+                                        return ctx.method().name().equalsIgnoreCase(p.getMethod().get().name())
+                                                && StringUtils.equals(ctx.path(), p.getPathPattern());
                                     }
 
                                     // Is Admin
                                     if (GeneralUtils.URL_PATH_SEPARATOR.equals(p.getPathPattern())) {
                                         // Nasty solution to a problem of how do we identify the call is to an admin's mock...
                                         // TODO This will improve when we update isInboundPathMultiUserPath to use a cache instead.
-                                        final String userCtxPathSegment = mockedRestServerEngineUtils.extractMultiUserCtxPathSegment(request.pathInfo());
+                                         final String userCtxPathSegment = mockedRestServerEngineUtils.extractMultiUserCtxPathSegment(ctx.path());
                                         return !mockedRestServerEngineUtils.isInboundPathMultiUserPath(userCtxPathSegment);
                                     }
 
-                                    return StringUtils.startsWith(request.pathInfo(), p.getPathPattern());
+                                    return StringUtils.startsWith(ctx.path(), p.getPathPattern());
                                 })
                         ) {
                             break;
@@ -405,7 +367,7 @@ public class MockedRestServerEngine {
                     // Could be no amendment is provided (in which case this request will default to the original response)
                     if (responseAmendmentOpt.isPresent()) {
 
-                        final String body = amendResponse(responseAmendmentOpt, response);
+                        final String body = amendResponse(responseAmendmentOpt, ctx);
 
                         return Optional.of(body);
                     }
@@ -420,13 +382,13 @@ public class MockedRestServerEngine {
     }
 
     private String amendResponse(final Optional<LiveloggingUserOverrideResponse> responseAmendmentOpt,
-                               final Response response) {
+                               final Context ctx) {
 
         final LiveloggingUserOverrideResponse responseAmendment = responseAmendmentOpt.get();
 
         if (!responseAmendment.getResponseHeaders().isEmpty()) {
 
-            final HttpServletResponse httpServletResponse = response.raw();
+            final HttpServletResponse httpServletResponse = ctx.res();
 
             responseAmendment
                 .getResponseHeaders()
@@ -451,62 +413,52 @@ public class MockedRestServerEngine {
 
         }
 
-        response.status(responseAmendment.getStatus());
-        response.body(responseAmendment.getBody());
+        ctx.status(responseAmendment.getStatus());
+        ctx.result(responseAmendment.getBody());
 
         return responseAmendment.getBody();
     }
 
-    private String handleNotFoundResponse(final Response response) {
-
-        response.status(HttpStatus.NOT_FOUND.value());
-
+    private String handleNotFoundResponse(final Context ctx) {
+        ctx.status(HttpStatus.NOT_FOUND.value());
         return "";
     }
 
-    private boolean isWebSocketUpgradeRequest(final Request request) {
+    private boolean isWebSocketUpgradeRequest(final Context ctx) {
 
-        final Set<String> headerNames = request.headers();
+        final Map<String, String> headers = ctx.headerMap();
 
-        return headerNames.contains(HttpHeaders.UPGRADE)
-                && headerNames.contains("Sec-WebSocket-Key")
-                && "websocket".equalsIgnoreCase(request.headers(HttpHeaders.UPGRADE));
+        return headers.containsKey(HttpHeaders.UPGRADE)
+                && headers.containsKey("Sec-WebSocket-Key")
+                && "websocket".equalsIgnoreCase(headers.get(HttpHeaders.UPGRADE));
     }
 
-    boolean blockLoggingResponse(final Request request,
-                          final Response response,
+    boolean blockLoggingResponse(final Context ctx,
                           final boolean proxyMode) {
-
         if (logger.isDebugEnabled()) {
             logger.debug("liveBlockEnabled: " + liveBlockingModeEnabled.get());
-            logger.debug("inbound method: " + request.requestMethod());
-            logger.debug("inbound path: " + request.pathInfo());
+            logger.debug("inbound method: {}" , ctx.method());
+            logger.debug("inbound path: {}" , ctx.path());
         }
-
         if (this.liveBlockingModeEnabled.get()) {
-
             // Check if this call matches any endpoints that should be blocked...
             if (liveBlockPathsRef.get()
                     .stream()
                     .anyMatch(p ->
-                            p.getMethod().name().equalsIgnoreCase(request.requestMethod())
-                                    && GeneralUtils.matchPaths(p.getPath(), request.pathInfo()))) {
-
+                            p.getMethod().name().equalsIgnoreCase(ctx.method().name())
+                                    && GeneralUtils.matchPaths(p.getPath(), ctx.path()))) {
                 // If so then send response details to live logging console via WS...
                 liveLoggingHandler.broadcast(
                         LiveLoggingUtils.buildLiveLogInterceptedResponseDTO(
-                                request.attribute(GeneralUtils.LOG_REQ_ID),
-                                request.pathInfo(),
-                                response.raw().getStatus(),
-                                mockedRestServerEngineUtils.extractResponseHeadersAsMap(response),
-                                response.body(),
+                                ctx.attribute(GeneralUtils.LOG_REQ_ID),
+                                ctx.path(),
+                                ctx.status().getCode(),
+                                mockedRestServerEngineUtils.extractResponseHeadersAsMap(ctx),
+                                ctx.result(),
                                 proxyMode));
-
                 return true;
             }
-
         }
-
         return false;
     }
 
@@ -517,37 +469,6 @@ public class MockedRestServerEngine {
         proxyService.clearAllSessions();
         mockOrderingCounterService.clearState();
         serverSideEventService.clearState();
-
-    }
-
-    void handleCORS(final MockedServerConfigDTO config) {
-
-        final String enableCors = config.getNativeProperties().get(GeneralUtils.ENABLE_CORS_PARAM);
-
-        if (!Boolean.TRUE.toString().equalsIgnoreCase(enableCors)) {
-            return;
-        }
-
-        Spark.options(SPARK_WILDCARD_PATH, (request, response) -> {
-
-            final String accessControlRequestHeaders = request.headers(HttpHeaders.ACCESS_CONTROL_REQUEST_HEADERS);
-
-            if (accessControlRequestHeaders != null) {
-                response.header(HttpHeaders.ACCESS_CONTROL_ALLOW_HEADERS, accessControlRequestHeaders);
-            }
-
-            final String accessControlRequestMethod = request.headers(HttpHeaders.ACCESS_CONTROL_REQUEST_METHOD);
-
-            if (accessControlRequestMethod != null) {
-
-                response.header(HttpHeaders.ACCESS_CONTROL_ALLOW_METHODS, accessControlRequestMethod);
-            }
-
-            return HttpStatus.OK.name();
-        });
-
-        Spark.before((request, response) ->
-            response.header(HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN, GeneralUtils.PATH_WILDCARD));
 
     }
 
