@@ -2,7 +2,7 @@ package com.smockin.mockserver.service;
 
 import com.smockin.admin.persistence.entity.RestfulMock;
 import com.smockin.mockserver.service.dto.RestfulResponseDTO;
-import spark.Request;
+import io.javalin.http.Context;
 
 public interface JavaScriptResponseHandler {
 
@@ -46,6 +46,6 @@ public interface JavaScriptResponseHandler {
             + " response;"
             + "}";
 
-    RestfulResponseDTO executeUserResponse(final Request req, final RestfulMock mock);
+    RestfulResponseDTO executeUserResponse(final Context ctx, final RestfulMock mock);
 
 }
