@@ -6,17 +6,17 @@ import com.smockin.admin.service.SmockinUserService;
 import com.smockin.admin.service.UserKeyValueDataService;
 import com.smockin.mockserver.service.dto.RestfulResponseDTO;
 import com.smockin.utils.GeneralUtils;
-import jdk.nashorn.api.scripting.NashornScriptEngineFactory;
-import jdk.nashorn.api.scripting.ScriptObjectMirror;
+import io.javalin.http.Context;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.client.utils.URLEncodedUtils;
+import org.openjdk.nashorn.api.scripting.NashornScriptEngineFactory;
+import org.openjdk.nashorn.api.scripting.ScriptObjectMirror;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import spark.Request;
 
 import javax.script.Bindings;
 import javax.script.ScriptContext;
@@ -45,7 +45,7 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
     private final static String CARRIAGE_RETURN_REGEX = "\\r\\n|\\r|\\n";
     private final String extensionsDir = "js-extensions/";
 
-    public RestfulResponseDTO executeUserResponse(final Request req, final RestfulMock mock) {
+    public RestfulResponseDTO executeUserResponse(final Context ctx, final RestfulMock mock) {
         logger.debug("executeUserResponse called");
 
         Object engineResponse;
@@ -54,8 +54,8 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
 
             engineResponse = executeJS(
                     defaultRequestObject
-                        + populateRequestObjectWithInbound(req, mock.getPath(), mock.getCreatedBy().getCtxPath())
-                        + populateKVPs(req, mock)
+                        + populateRequestObjectWithInbound(ctx, mock.getPath(), mock.getCreatedBy().getCtxPath())
+                        + populateKVPs(ctx, mock)
                         + keyValuePairFindFunc
                         + defaultResponseObject
                         + userResponseFunctionInvoker
@@ -89,28 +89,25 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
         return buildEngine().eval(js);
     }
 
-    String populateRequestObjectWithInbound(final Request req, final String mockPath, final String ctxPath) {
+    String populateRequestObjectWithInbound(final Context ctx, final String mockPath, final String ctxPath) {
 
-        final Map<String, String> reqHeaders =
-                req.headers()
-                    .stream()
-                    .collect(Collectors.toMap(k -> k, k -> req.headers(k)));
+        final Map<String, String> reqHeaders = ctx.headerMap();
 
         final StringBuilder reqObject = new StringBuilder();
 
         reqObject.append("request.path=")
-                .append("'").append(req.pathInfo()).append("'")
+                .append("'").append(ctx.path()).append("'")
                 .append("; ");
 
-        if (StringUtils.isNotBlank(req.body())) {
+        if (StringUtils.isNotBlank(ctx.body())) {
             reqObject.append("request.body=")
-                    .append("'").append(removeLineBreaks(req.body())).append("'")
+                    .append("'").append(removeLineBreaks(ctx.body())).append("'")
                     .append(";");
         }
 
-        final String sanitizedInboundPath = GeneralUtils.sanitizeMultiUserPath(smockinUserService.getUserMode(), req.pathInfo(), ctxPath);
+        final String sanitizedInboundPath = GeneralUtils.sanitizeMultiUserPath(smockinUserService.getUserMode(), ctx.path(), ctxPath);
         applyMapValuesToStringBuilder("request.pathVars", GeneralUtils.findAllPathVars(sanitizedInboundPath, mockPath), reqObject);
-        applyMapValuesToStringBuilder("request.parameters", extractAllRequestParams(req), reqObject);
+        applyMapValuesToStringBuilder("request.parameters", extractAllRequestParams(ctx), reqObject);
         applyMapValuesToStringBuilder("request.headers", reqHeaders, reqObject);
 
         return reqObject.toString();
@@ -131,22 +128,23 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
                         .append(";"));
     }
 
-    Map<String, String> extractAllRequestParams(final Request req) {
+    Map<String, String> extractAllRequestParams(final Context ctx) {
 
-        // Java Spark does not provide a convenient way of extracting form based request parameters,
+        // Javalin does not provide a convenient way of extracting form based request parameters,
         // so have to parse these manually.
-        if (req.contentType() != null
-                && (req.contentType().contains(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-                ||  req.contentType().contains(MediaType.MULTIPART_FORM_DATA_VALUE))) {
+        if (ctx.contentType() != null
+                && (ctx.contentType().contains(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+                ||  ctx.contentType().contains(MediaType.MULTIPART_FORM_DATA_VALUE))) {
 
-            return URLEncodedUtils.parse(req.body(), Charset.defaultCharset())
+            return URLEncodedUtils.parse(ctx.body(), Charset.defaultCharset())
                     .stream()
                     .collect(HashMap::new, (m, v) -> m.put(v.getName(), v.getValue()), HashMap::putAll);
         }
 
-        return req.queryParams()
+        return ctx.queryParamMap()
+                .entrySet()
                 .stream()
-                .collect(Collectors.toMap(k -> k, k -> req.queryParams(k)));
+                .collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().isEmpty() ? "" : e.getValue().get(0)));
     }
 
     Set<Map.Entry<String, String>> convertResponseHeaders(final ScriptObjectMirror response) {
@@ -165,7 +163,7 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
         return responseHeaders.entrySet();
     }
 
-    String populateKVPs(final Request req, final RestfulMock mock) throws ScriptException {
+    String populateKVPs(final Context ctx, final RestfulMock mock) throws ScriptException {
         logger.debug("populateKVPs called");
 
         final String handleResponseFunc = GeneralUtils.removeJsComments(mock.getJavaScriptHandler().getSyntax());
@@ -186,7 +184,7 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
             }
 
             final int closingParenthesisPos = StringUtils.indexOf(handleResponseFunc, ")", startPos);
-            final String sanitizedKey = findKvpKey(startPos, closingParenthesisPos, req, mock, keyValuePairFuncPrefix, handleResponseFunc);
+            final String sanitizedKey = findKvpKey(startPos, closingParenthesisPos, ctx, mock, keyValuePairFuncPrefix, handleResponseFunc);
 
             if (sanitizedKey != null) {
                 final UserKeyValueDataDTO userKeyValueDataDTO = userKeyValueDataService.loadByKey(sanitizedKey, mockOwnerUserId);
@@ -205,7 +203,7 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
         return defaultKeyValuePairStoreObject;
     }
 
-    private String findKvpKey(final int startPos, final int closingParenthesisPos, final Request req, final RestfulMock mock, final String keyValuePairFuncPrefix, final String handleResponseFunc)
+    private String findKvpKey(final int startPos, final int closingParenthesisPos, final Context ctx, final RestfulMock mock, final String keyValuePairFuncPrefix, final String handleResponseFunc)
             throws ScriptException {
         logger.debug("findKvpKey called");
 
@@ -236,32 +234,30 @@ public class JavaScriptResponseHandlerImpl implements JavaScriptResponseHandler 
             if (requestObjectField.startsWith("pathVars")) {
 
                 final String pathVarsObjectField = StringUtils.remove(requestObjectField, "pathVars").trim();
-                final String sanitizedInboundPath = GeneralUtils.sanitizeMultiUserPath(smockinUserService.getUserMode(), req.pathInfo(), mock.getCreatedBy().getCtxPath());
+                final String sanitizedInboundPath = GeneralUtils.sanitizeMultiUserPath(smockinUserService.getUserMode(), ctx.path(), mock.getCreatedBy().getCtxPath());
                 sanitizedKey = GeneralUtils.findAllPathVars(sanitizedInboundPath, mock.getPath())
                         .get(extractObjectField(StringUtils.lowerCase(pathVarsObjectField)));
 
             } else if ("body".equals(requestObjectField)) {
 
-                if (StringUtils.isBlank(req.body())) {
+                if (StringUtils.isBlank(ctx.body())) {
                     throw new ScriptException(invalidMsgPrefix + "request.body is undefined");
                 }
 
-                sanitizedKey = removeLineBreaks(req.body());
+                sanitizedKey = removeLineBreaks(ctx.body());
 
             } else if (requestObjectField.startsWith("headers")) {
 
                 final String headersObjectField = StringUtils.remove(requestObjectField, "headers").trim();
 
-                sanitizedKey = req.headers()
-                        .stream()
-                        .collect(Collectors.toMap(k -> k, k -> req.headers(k)))
+                sanitizedKey = ctx.headerMap()
                         .get(extractObjectField(headersObjectField));
 
             } else if (requestObjectField.startsWith("parameters")) {
 
                 final String parametersObjectField = StringUtils.remove(requestObjectField, "parameters").trim();
 
-                sanitizedKey = extractAllRequestParams(req).get(extractObjectField(parametersObjectField));
+                sanitizedKey = extractAllRequestParams(ctx).get(extractObjectField(parametersObjectField));
 
             } else {
                 throw new ScriptException(invalidMsgPrefix + "Unable to determine request based key look up");
