@@ -2,10 +2,10 @@ package com.smockin.mockserver.service;
 
 import com.smockin.admin.exception.RecordNotFoundException;
 import com.smockin.admin.exception.ValidationException;
-import com.smockin.mockserver.service.dto.SseMessageDTO;
 import com.smockin.mockserver.service.dto.PushClientDTO;
-import spark.Request;
-import spark.Response;
+import com.smockin.mockserver.service.dto.SseMessageDTO;
+import io.javalin.http.Context;
+
 import java.io.IOException;
 import java.util.List;
 
@@ -16,7 +16,7 @@ public interface ServerSideEventService {
 
     String SSE_EVENT_STREAM_HEADER = "text/event-stream;charset=UTF-8";
 
-    void register(final String path, final long heartBeatMillis, final boolean proxyPushIdOnConnect, final Request request, final Response response) throws IOException;
+    void register(final String path, final long heartBeatMillis, final boolean proxyPushIdOnConnect, final Context ctx) throws IOException;
     List<PushClientDTO> getClientConnections(final String mockExtId, final String token) throws RecordNotFoundException, ValidationException;
     void addMessage(final String id, final SseMessageDTO dto);
     void interruptAndClearAllHeartBeatThreads();
