@@ -4,11 +4,11 @@ import com.smockin.admin.exception.RecordNotFoundException;
 import com.smockin.admin.exception.ValidationException;
 import com.smockin.admin.persistence.entity.RestfulMock;
 import com.smockin.mockserver.service.dto.RestfulResponseDTO;
-import spark.Request;
+import io.javalin.http.Context;
 
 public interface StatefulService {
 
-    RestfulResponseDTO process(final Request req, final RestfulMock mock);
+    RestfulResponseDTO process(final Context ctx, final RestfulMock mock);
     void resetState(final String externalId, final String userToken) throws RecordNotFoundException, ValidationException;
 
 }
