@@ -2,10 +2,11 @@ package com.smockin.mockserver.service.ws;
 
 import com.smockin.mockserver.service.WebSocketService;
 import com.smockin.mockserver.service.enums.WebSocketCommandEnum;
+import org.eclipse.jetty.websocket.api.Callback;
 import org.eclipse.jetty.websocket.api.Session;
 import org.eclipse.jetty.websocket.api.annotations.OnWebSocketClose;
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketConnect;
 import org.eclipse.jetty.websocket.api.annotations.OnWebSocketMessage;
+import org.eclipse.jetty.websocket.api.annotations.OnWebSocketOpen;
 import org.eclipse.jetty.websocket.api.annotations.WebSocket;
 
 import java.io.IOException;
@@ -24,7 +25,7 @@ public class SparkWebSocketEchoService {
         this.isMultiUserMode = isMultiUserMode;
     }
 
-    @OnWebSocketConnect
+    @OnWebSocketOpen
     public void connected(final Session session) {
         webSocketService.registerSession(session, isMultiUserMode);
     }
@@ -38,7 +39,7 @@ public class SparkWebSocketEchoService {
     public void message(final Session session, final String message) throws IOException {
 
         if (WebSocketCommandEnum.SMOCKIN_ID.name().equals(message)) {
-            session.getRemote().sendString(webSocketService.getExternalId(session));
+            session.sendText(webSocketService.getExternalId(session), Callback.NOOP);
             return;
         }
 
